+++ CHANGELOG.md
@@ -24,5 +24,6 @@ The minor version will be incremented upon a breaking change and the patch versi
 - api: optimize getLeaderSchedule ([#9](https://github.com/solana-stream-solutions/solfees/pull/9))
 - geyser: wait all transactions before process block ([#10](https://github.com/solana-stream-solutions/solfees/pull/10))
 - frontend: init ([#8](https://github.com/solana-stream-solutions/solfees/pull/8))
+- geyser: use process_compute_budget_instructions ([#15](https://github.com/solana-stream-solutions/solfees/pull/15))
 
 ### Breaking